+++ modules/vorbis/audio_stream_ogg_vorbis.cpp
@@ -157,10 +157,10 @@ int AudioStreamPlaybackOggVorbis::_mix_frames_vorbis(AudioFrame *p_buffer, int p
 		}
 
 		err = vorbis_synthesis(&block, packet);
-		ERR_FAIL_COND_V_MSG(err != 0, p_frames, "Error during vorbis synthesis " + itos(err));
+		ERR_FAIL_COND_V_MSG(err != 0, p_frames, "Error during vorbis synthesis " + itos((int)err));
 
 		err = vorbis_synthesis_blockin(&dsp_state, &block);
-		ERR_FAIL_COND_V_MSG(err != 0, p_frames, "Error during vorbis block processing " + itos(err));
+		ERR_FAIL_COND_V_MSG(err != 0, p_frames, "Error during vorbis block processing " + itos((int)err));
 
 		have_packets_left = !packet->e_o_s;
 	}
@@ -324,30 +324,30 @@ void AudioStreamPlaybackOggVorbis::seek(double p_time) {
 
 			err = vorbis_synthesis(&block, packet);
 			if (err != OV_ENOTAUDIO) {
-				ERR_FAIL_COND_MSG(err != 0, "Error during vorbis synthesis " + itos(err) + ".");
+				ERR_FAIL_COND_MSG(err != 0, "Error during vorbis synthesis " + itos((int)err) + ".");
 
 				err = vorbis_synthesis_blockin(&dsp_state, &block);
-				ERR_FAIL_COND_MSG(err != 0, "Error during vorbis block processing " + itos(err) + ".");
+				ERR_FAIL_COND_MSG(err != 0, "Error during vorbis block processing " + itos((int)err) + ".");
 
 				int samples_out = vorbis_synthesis_pcmout(&dsp_state, nullptr);
 
 				if (granule_pos < 0) {
 					// We don't know where we are yet, so just keep on decoding.
 					err = vorbis_synthesis_read(&dsp_state, samples_out);
-					ERR_FAIL_COND_MSG(err != 0, "Error during vorbis read updating " + itos(err) + ".");
+					ERR_FAIL_COND_MSG(err != 0, "Error during vorbis read updating " + itos((int)err) + ".");
 				} else if (granule_pos + samples_out >= desired_sample) {
 					// Our sample is in this block. Skip the beginning of the block up to the sample, then
 					// return.
 					int skip_samples = (int)(desired_sample - granule_pos);
 					err = vorbis_synthesis_read(&dsp_state, skip_samples);
-					ERR_FAIL_COND_MSG(err != 0, "Error during vorbis read updating " + itos(err) + ".");
+					ERR_FAIL_COND_MSG(err != 0, "Error during vorbis read updating " + itos((int)err) + ".");
 					have_samples_left = skip_samples < samples_out;
 					have_packets_left = !packet->e_o_s;
 					return;
 				} else {
 					// Our sample is not in this block. Skip it.
 					err = vorbis_synthesis_read(&dsp_state, samples_out);
-					ERR_FAIL_COND_MSG(err != 0, "Error during vorbis read updating " + itos(err) + ".");
+					ERR_FAIL_COND_MSG(err != 0, "Error during vorbis read updating " + itos((int)err) + ".");
 					granule_pos += samples_out;
 				}
 			}
@@ -438,7 +438,7 @@ void AudioStreamOggVorbis::maybe_update_info() {
 		}
 
 		err = vorbis_synthesis_headerin(&info, &comment, packet);
-		ERR_FAIL_COND_MSG(err != 0, "Error parsing header packet " + itos(i) + ": " + itos(err));
+		ERR_FAIL_COND_MSG(err != 0, "Error parsing header packet " + itos(i) + ": " + itos((int)err));
 	}
 
 	packet_sequence->set_sampling_rate(info.rate);
